+++ lib/components/ui/multi-select.tsx
@@ -264,7 +264,7 @@ export const MultiSelect = React.forwardRef<
                         }
                       }}
                       onKeyDown={handleKeyDown}
-                      className="focus:lui-outline-none disabled:lui-bg-transparent disabled:placeholder:lui-text-ocean-light-40"
+                      className="lui-caret-ocean-primary-10 focus:lui-outline-none disabled:lui-bg-transparent disabled:placeholder:lui-text-ocean-light-40"
                       ref={inputRef}
                       disabled={props.disabled}
                     />
@@ -279,7 +279,7 @@ export const MultiSelect = React.forwardRef<
               </button>
             </PopoverTrigger>
             <PopoverContent
-              className="lui-w-auto lui-min-w-[--radix-popover-trigger-width] lui-max-w-[--radix-popover-trigger-width] lui-p-0 lui-font-bca"
+              className="lui-min-w-[--radix-popover-trigger-width] lui-max-w-[--radix-popover-trigger-width]"
               align="start"
               onEscapeKeyDown={() => setIsPopoverOpen(false)}
               onOpenAutoFocus={(e) => e.preventDefault()}