+++ lib/components/ui/select.tsx
@@ -177,7 +177,7 @@ export const Select = React.forwardRef<HTMLButtonElement, SelectProps>(
                 )}
                 {...props}
               >
-                <div className="lui-mx-auto lui-flex lui-w-full lui-items-center lui-justify-between">
+                <div className="lui-mx-auto lui-flex lui-w-full lui-items-center lui-justify-between lui-caret-ocean-primary-10">
                   <input
                     type="text"
                     placeholder={placeholder}
@@ -225,7 +225,7 @@ export const Select = React.forwardRef<HTMLButtonElement, SelectProps>(
               </button>
             </PopoverTrigger>
             <PopoverContent
-              className="lui-w-auto lui-min-w-[--radix-popover-trigger-width] lui-max-w-[--radix-popover-trigger-width] lui-p-0 lui-font-bca"
+              className="lui-min-w-[--radix-popover-trigger-width] lui-max-w-[--radix-popover-trigger-width] lui-font-bca"
               align="start"
               onEscapeKeyDown={() => setIsPopoverOpen(false)}
               onOpenAutoFocus={(e) => e.preventDefault()}