+++ lib/components/ui/tabs.tsx
@@ -4,10 +4,11 @@ import * as TabsPrimitive from "@radix-ui/react-tabs";
 import { cn } from "@/lib/utils";
 
 type TabsProps = React.ComponentPropsWithoutRef<typeof TabsPrimitive.Root> & {
-  list?: { className?: string };
-  trigger?: { className?: string };
+  listProps?: Partial<TabsListProps>;
+  triggerProps?: Partial<TabsTriggerProps>;
   tabMenus: TabMenu[];
-  setTabMenus?: React.Dispatch<React.SetStateAction<TabMenu[]>>;
+  selectedTab?: string;
+  setSelectedTab?: React.Dispatch<React.SetStateAction<string>>;
   tabsStyle?: "underline" | "filled";
   size?: "medium" | "large";
 };
@@ -19,25 +20,32 @@ const Tabs = React.forwardRef<
   (
     {
       children,
-      list,
-      trigger,
+      listProps,
+      triggerProps,
       tabMenus,
       tabsStyle = "underline",
       size = "medium",
+      selectedTab,
+      setSelectedTab,
       ...props
     },
     ref,
   ) => (
-    <TabsPrimitive.Root ref={ref} {...props}>
-      <TabsList tabsStyle={tabsStyle} size={size} className={list?.className}>
+    <TabsPrimitive.Root
+      ref={ref}
+      {...props}
+      value={selectedTab}
+      onValueChange={setSelectedTab}
+    >
+      <TabsList tabsStyle={tabsStyle} size={size} {...listProps}>
         {tabMenus.map((tabMenu) => (
           <TabsTrigger
             key={tabMenu.value}
             tabsStyle={tabsStyle}
             size={size}
-            className={trigger?.className}
             value={tabMenu.value}
             disabled={tabMenu.disabled}
+            {...triggerProps}
           >
             {tabMenu.name}
           </TabsTrigger>
@@ -119,18 +127,10 @@ const TabsTrigger = React.forwardRef<
 ));
 TabsTrigger.displayName = TabsPrimitive.Trigger.displayName;
 
-const TabsContent = React.forwardRef<
-  React.ElementRef<typeof TabsPrimitive.Content>,
-  React.ComponentPropsWithoutRef<typeof TabsPrimitive.Content>
->(({ className, ...props }, ref) => (
-  <TabsPrimitive.Content ref={ref} className={cn(className)} {...props} />
-));
-TabsContent.displayName = TabsPrimitive.Content.displayName;
-
 export type TabMenu = {
   name: string;
   value: string;
   disabled?: boolean;
 };
 
-export { Tabs, TabsList, TabsTrigger, TabsContent };
+export { Tabs };