+++ lib/hooks/useWindowSize.ts
@@ -1,35 +1,37 @@
 import { useState, useEffect } from 'react';
 
 interface WindowSize {
-    width: number;
-    height: number;
+    width: number | undefined;
+    height: number | undefined;
 }
 
 function useWindowSize(): WindowSize {
     const [windowSize, setWindowSize] = useState<WindowSize>({
-        width: window.innerWidth,
-        height: window.innerHeight,
+        width: undefined,
+        height: undefined,
     });
 
     useEffect(() => {
-        function handleResize() {
-            setWindowSize({
-                width: window.innerWidth,
-                height: window.innerHeight,
-            });
+        // Check if window is defined (client-side)
+        if (typeof window !== 'undefined') {
+            function handleResize() {
+                setWindowSize({
+                    width: window.innerWidth,
+                    height: window.innerHeight,
+                });
+            }
+
+            window.addEventListener('resize', handleResize);
+
+            // Call handler right away so state gets updated with initial window size
+            handleResize();
+
+            // Remove event listener on cleanup
+            return () => window.removeEventListener('resize', handleResize);
         }
-
-        window.addEventListener('resize', handleResize);
-
-        // Call handler right away so state gets updated with initial window size
-        handleResize();
-
-        // Remove event listener on cleanup
-        return () => window.removeEventListener('resize', handleResize);
     }, []); // Empty array ensures that effect is only run on mount and unmount
 
     return windowSize;
 }
 
 export default useWindowSize;
-