+++ lib/index.ts
@@ -149,7 +149,7 @@ export {
     TableHeader,
     TableRow,
 } from "./components/ui/table";
-export { Tabs, TabsContent } from "./components/ui/tabs";
+export { Tabs } from "./components/ui/tabs";
 export type { TabMenu } from "./components/ui/tabs";
 export {
     Dialog,
@@ -168,20 +168,23 @@ export { PaginationDetailed } from "./components/ui/pagination-detailed";
 export { Stepper } from "./components/ui/stepper";
 export { StepperCircle } from "./components/ui/stepper-circle";
 export { addMinutes, TimePicker } from "@/components/ui/time-picker";
+export type { ValidTime } from "@/components/ui/time-picker";
 
 //NOTE - v0.4.0
 export { PlusOutline } from "./components/ui/icon/PlusOutline";
 export { MinusOutline } from "./components/ui/icon/MinusOutline";
 export { ArrowRightLargeOutline } from "./components/ui/icon/ArrowRightLargeOutline";
 export { LightbulbOutline } from "./components/ui/icon/LightbulbOutline";
+export { MediaPlayFilled } from "./components/ui/icon/MediaPlayFilled";
+export { FilePDFFilled } from "./components/ui/icon/FilePDFFilled";
 export { Accordion, AccordionItem, AccordionTrigger, AccordionContent } from "./components/ui/accordion";
 
 //NOTE - Beta
-// export { Calendar } from "./components/ui/calendar";
-// export {
-//     DropdownMenu,
-//     DropdownMenuContent,
-//     DropdownMenuItem,
-//     DropdownMenuTrigger,
-// } from "./components/ui/dropdown-menu";
-// export { Popover, PopoverContent, PopoverTrigger } from "./components/ui/popover";
+export { Calendar } from "./components/ui/calendar";
+export {
+    DropdownMenu,
+    DropdownMenuContent,
+    DropdownMenuItem,
+    DropdownMenuTrigger,
+} from "./components/ui/dropdown-menu";
+export { Popover, PopoverContent, PopoverTrigger } from "./components/ui/popover";